Separate AutoUpdate from InitializeListener
